[feat/3] refactor: html 생성, 추가 관련 로직 분리
